add: home page adition and folder restructured
