+++ styles/Home.module.css
@@ -1,110 +1,1 @@
-.sectionTitle {
-	background-color: var(--color-bg-light);
-	height: 10rem;
-	margin-top: 0;
-	position: relative;
-}
 
-.titleLine {
-	height: 0.8rem;
-	background-color: var(--color-primary-light);
-	opacity: 70%;
-	width: 100%;
-	position: absolute;
-	top: 60%;
-}
-
-.wrapper {
-	position: relative;
-	display: flex;
-	justify-content: center;
-
-	margin-left: 10vw;
-	margin-right: 10vw;
-}
-
-.container {
-	max-width: 1100px;
-
-	margin: 0;
-	padding: 0;
-	display: grid;
-	grid-template-areas:
-		'image1 image1 text1'
-		'text2 image2 image2';
-	grid-template-columns: 45% auto 45%;
-	grid-column-gap: 3rem;
-	grid-row-gap: 3rem;
-
-	margin-bottom: 4rem;
-}
-
-.image1 {
-	grid-area: image1;
-	position: relative;
-	top: -1.5rem;
-	overflow: hidden;
-	background: url('../public/tornado.jpg');
-	background-position: center;
-	background-size: cover;
-}
-
-.image2 {
-	grid-area: image2;
-	display: flex;
-	align-items: center;
-	overflow: hidden;
-	background: url('../public/interbatuc3.jpg');
-	background-position: center;
-	background-size: cover;
-}
-
-.text1 {
-	grid-area: text1;
-	position: relative;
-	top: -3.5rem;
-}
-
-.text2 {
-	grid-area: text2;
-	margin-top: -5rem;
-}
-
-.text1 {
-	margin-top: -6.5rem;
-}
-
-.text1 p,
-.text2 p {
-	font-size: 1.6rem;
-	text-align: justify;
-	line-height: 150%;
-	text-indent: 15px;
-}
-
-.line {
-	height: 0.6rem;
-	background-color: var(--color-primary-light);
-	opacity: 70%;
-	position: relative;
-	top: 6.4rem;
-	width: 120%;
-	left: calc(-1 * (100vw - 100% / 0.45) / 2);
-}
-
-@media (max-width: 768px) {
-	.container {
-		grid-template-columns: 1fr;
-		grid-template-areas:
-			'text1'
-			'image1'
-			'text2'
-			'image2';
-	}
-
-	.image1,
-	.image2 {
-		width: 100%;
-		height: 40vh;
-	}
-}
